Style: Rename & Style event sample
