--- app/src/main/java/com/jbrunoo/seoul_go/data/dataSource/local/room/EventDao.kt
@@ -1,4 +1,0 @@
-package com.jbrunoo.seoul_go.data.dataSource.local.room
-
-interface EventDao {
-}
