+++ app/src/main/java/com/jbrunoo/seoul_go/data/dataSource/local/room/event/EventDao.kt
@@ -1,0 +1,4 @@
+package com.jbrunoo.seoul_go.data.dataSource.local.room.event
+
+interface EventDao {
+}
