+++ app/src/main/java/com/jbrunoo/seoul_go/data/dataSource/local/room/event/EventLocalDataSource.kt
@@ -1,0 +1,5 @@
+package com.jbrunoo.seoul_go.data.dataSource.local.room.event
+
+interface EventLocalDataSource {
+
+}
