+++ app/src/main/java/com/jbrunoo/seoul_go/data/dataSource/local/room/event/EventLocalDataSourceImpl.kt
@@ -1,0 +1,4 @@
+package com.jbrunoo.seoul_go.data.dataSource.local.room.event
+
+class EventLocalDataSourceImpl: EventLocalDataSource {
+}
