+++ app/src/main/java/com/jbrunoo/seoul_go/data/dataSource/remote/EventRemoteDataSource.kt
@@ -4,5 +4,5 @@ import com.jbrunoo.seoul_go.data.dataSource.remote.dto.EventResponse
 import kotlinx.coroutines.flow.Flow
 
 interface EventRemoteDataSource {
-    fun getEvents() : Flow<List<EventResponse>>
+    fun fetchEvents() : Flow<List<EventResponse>>
 }
