+++ app/src/main/java/com/jbrunoo/seoul_go/data/dataSource/remote/EventRemoteDataSourceImpl.kt
@@ -8,7 +8,7 @@ import javax.inject.Inject
 
 class EventRemoteDataSourceImpl @Inject constructor(private val eventService: EventService) :
     EventRemoteDataSource {
-    override fun getEvents(): Flow<List<EventResponse>> {
+    override fun fetchEvents(): Flow<List<EventResponse>> {
         return flow {
             emit(eventService.getEvents())
         }
