+++ app/src/main/java/com/jbrunoo/seoul_go/data/dataSource/remote/api/EventService.kt
@@ -1,8 +1,6 @@
 package com.jbrunoo.seoul_go.data.dataSource.remote.api
 
 import com.jbrunoo.seoul_go.data.dataSource.remote.dto.EventResponse
-import com.jbrunoo.seoul_go.domain.entity.Event
-import retrofit2.Call
 import retrofit2.http.GET
 
 interface EventService {
