+++ app/src/main/java/com/jbrunoo/seoul_go/data/repository/EventRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.jbrunoo.seoul_go.data.repository
 
 import com.jbrunoo.seoul_go.data.dataSource.remote.EventRemoteDataSource
 import com.jbrunoo.seoul_go.data.dataSource.remote.dto.toEvent
-import com.jbrunoo.seoul_go.domain.entity.Event
+import com.jbrunoo.seoul_go.domain.model.Event
 import com.jbrunoo.seoul_go.domain.repository.EventRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
@@ -10,7 +10,7 @@ import javax.inject.Inject
 
 class EventRepositoryImpl @Inject constructor(private val eventRemoteDataSource: EventRemoteDataSource) :
     EventRepository {
-    override fun getEvents(): Flow<List<Event>> =
-        eventRemoteDataSource.getEvents().map { it.map { it.toEvent() } }
+    override fun fetchEvents(): Flow<List<Event>> =
+        eventRemoteDataSource.fetchEvents().map { it.map { it.toEvent() } }
 }
