+++ app/src/main/java/com/jbrunoo/seoul_go/domain/model/Event.kt
@@ -1,0 +1,8 @@
+package com.jbrunoo.seoul_go.domain.model
+
+data class Event(
+    val id: Int,
+    val name: String,
+    val image: String,
+    val isFavorite: Boolean,
+)
