--- app/src/main/java/com/jbrunoo/seoul_go/domain/useCase/GetLikedEventUseCase.kt
@@ -1,4 +1,0 @@
-package com.jbrunoo.seoul_go.domain.useCase
-
-class GetLikedEventUseCase {
-}
