--- app/src/main/java/com/jbrunoo/seoul_go/domain/useCase/LikeEventUseCase.kt
@@ -1,4 +1,0 @@
-package com.jbrunoo.seoul_go.domain.useCase
-
-class LikeEventUseCase {
-}
