+++ app/src/main/java/com/jbrunoo/seoul_go/domain/useCase/event/GetLikedEventUseCase.kt
@@ -1,0 +1,4 @@
+package com.jbrunoo.seoul_go.domain.useCase.event
+
+class GetLikedEventUseCase {
+}
